Add support for generic fold output transforms
This commit adds support for folding the output transforms
of gemm-based operator should there be view like ops
after the gemm-based operators.

Additionally folding of view-like ops is lifted to
FoldTranspose pass from AlignTiling.
