+++ mlir/lib/Dialect/Rock/Transforms/AlignTiling.cpp
@@ -60,199 +60,6 @@ struct RockLinalgAlignPass
   void runOnOperation() override;
 };
 
-// This rewrite will rewrite the linalg IO that has view like-ops surrounding
-// them to be consumed by the linalg operation itself adjusting the indexing
-// maps to faithfully represent them.
-struct InlineViewLikeOperandsLinalgRewritePattern
-    : public OpRewritePattern<linalg::GenericOp> {
-  using OpRewritePattern<linalg::GenericOp>::OpRewritePattern;
-
-  LogicalResult matchAndRewrite(linalg::GenericOp laGeneric,
-                                PatternRewriter &rewriter) const override;
-};
-
-// This function will take strides calculated using reassociation of dimensions
-// that are modified by view-like ops and recompute the affine expr from src
-// view to dest view.
-static void getDelinearizedAffineExpr(mlir::ArrayRef<int64_t> strides,
-                                      mlir::ArrayRef<int64_t> shapes,
-                                      Builder &b, unsigned int position,
-                                      SmallVectorImpl<mlir::AffineExpr> &res) {
-  AffineExpr resultExpr = b.getAffineDimExpr(position);
-  int64_t rank = strides.size();
-  // If the rank is 1, expand or collapse shapes will just
-  // pass-through the dimensions.
-  if (rank == 1) {
-    res[0] = resultExpr;
-    return;
-  }
-  for (int i = 0; i < rank; i++) {
-    // If the current shape is 1 and the rank is non-zero,
-    // could only mean it is being broadcasted. Hence,
-    // putting zero.
-    if (shapes[i] == 1) {
-      res[i] = resultExpr * 0;
-    } else {
-      // Recording the vector offsets here.
-      res[i] = resultExpr;
-      // There is no point of putting a modulo if the size
-      // is equivalent to that.
-      if (i - 1 >= 0 && shapes[i] != strides[i - 1]) {
-        res[i] = res[i] % strides[i - 1];
-      }
-
-      if (shapes[i] > strides[i]) {
-        // We only need the floorDiv if the dimSize
-        // is larger than the stride
-        res[i] = res[i].floorDiv(strides[i]);
-      } else if (shapes[i] < strides[i]) {
-        // if the shape is smaller than the stride
-        // expr might as well be zero.
-        res[i] = res[i] * 0;
-      }
-
-      // The resultExpr has to propagated anyway for
-      // other dimensions where the recording in the above
-      // will do the neccesary checks to remove the modulo
-      if (i - 1 >= 0) {
-        resultExpr = resultExpr % strides[i - 1];
-      }
-      resultExpr = resultExpr.floorDiv(strides[i]);
-    }
-  }
-  return;
-}
-
-// This function will create a affine map that represent the mapping
-// from higher rank memref type to lower rank memref type.
-static AffineMap createHigherToLowerRankViewAffineMap(
-    PatternRewriter &rewriter,
-    ArrayRef<ReassociationIndices> reassociationIndices,
-    const MemRefType &higherRankType, const MemRefType &lowerRankType) {
-  SmallVector<AffineExpr, 4> resultExprs;
-  int iDimCount = 0;
-  for (const SmallVector<int64_t, 2> &groups : reassociationIndices) {
-    assert(!groups.empty() && "association indices groups cannot be empty");
-    unsigned groupSize = groups.size();
-    SmallVector<int64_t> suffixProduct(groupSize);
-    // Calculate suffix product for all collapse op source dimension sizes.
-    suffixProduct[groupSize - 1] = 1;
-    for (unsigned i = groupSize - 1; i > 0; i--)
-      suffixProduct[i - 1] =
-          suffixProduct[i] * higherRankType.getDimSize(groups[i]);
-    SmallVector<int64_t> shapes(groupSize);
-    for (unsigned i = 0; i < groupSize; i++) {
-      shapes[i] = higherRankType.getDimSize(groups[i]);
-    }
-    // Derive the index values along all dimensions of the source
-    // corresponding to the index wrt to collapsed shape op output.
-    SmallVector<AffineExpr, 4> srcIndexExpr(suffixProduct.size());
-    getDelinearizedAffineExpr(suffixProduct, shapes, rewriter, iDimCount++,
-                              srcIndexExpr);
-    for (unsigned i = 0; i < groupSize; i++) {
-      resultExprs.push_back(srcIndexExpr[i]);
-    }
-  }
-  auto representativeMap = AffineMap::get(
-      /*numDims=*/lowerRankType.cast<ShapedType>().getShape().size(),
-      /*numSymbols=*/0, resultExprs, rewriter.getContext());
-  return representativeMap;
-}
-
-// This function will traverse the operands of the linalg.generic folding
-// the view like ops to the indexing maps and returning the ultimate
-// folded map as well as the root operand.
-LogicalResult
-foldViewLikeOperands(PatternRewriter &rewriter, Value op, AffineMap &foldedMap,
-                     Value &rootOp,
-                     SmallVectorImpl<Operation *> &toBeErasedViewLikeOps) {
-  if (memref::CollapseShapeOp collapseOp =
-          op.getDefiningOp<memref::CollapseShapeOp>()) {
-    SmallVector<mlir::ReassociationIndices, 4U> reassociationIndices =
-        collapseOp.getReassociationIndices();
-    MemRefType lowerRankType = collapseOp.getType();
-    MemRefType higherRankType = collapseOp.getSrcType();
-    auto representativeMap = createHigherToLowerRankViewAffineMap(
-        rewriter, reassociationIndices, higherRankType, lowerRankType);
-    foldedMap = representativeMap.compose(foldedMap);
-    toBeErasedViewLikeOps.push_back(collapseOp);
-    return foldViewLikeOperands(rewriter, collapseOp.getViewSource(), foldedMap,
-                                rootOp, toBeErasedViewLikeOps);
-  }
-  if (memref::ExpandShapeOp expandOp =
-          op.getDefiningOp<memref::ExpandShapeOp>()) {
-    SmallVector<mlir::ReassociationIndices, 4U> reassociationIndices =
-        expandOp.getReassociationIndices();
-    MemRefType higherRankType = expandOp.getType();
-    MemRefType lowerRankType = expandOp.getSrcType();
-    auto representativeMap = createHigherToLowerRankViewAffineMap(
-        rewriter, reassociationIndices, higherRankType, lowerRankType);
-    // We take the inverse here because in expand shape it is going from lower
-    // to higher rank.
-    representativeMap = inversePermutation(representativeMap);
-    foldedMap = representativeMap.compose(foldedMap);
-    toBeErasedViewLikeOps.push_back(expandOp);
-    return foldViewLikeOperands(rewriter, expandOp.getViewSource(), foldedMap,
-                                rootOp, toBeErasedViewLikeOps);
-  }
-  rootOp = op;
-  return success();
-}
-
-LogicalResult InlineViewLikeOperandsLinalgRewritePattern::matchAndRewrite(
-    linalg::GenericOp laGeneric, PatternRewriter &rewriter) const {
-  Location loc = laGeneric.getLoc();
-  int ioCount = 0;
-
-  SmallVector<AffineMap, 4U> newMaps;
-  SmallVector<Value, 4U> newInputs;
-  unsigned int changedIOCount = 0;
-
-  SmallVector<Operation *> toBeErasedViewLikeOps;
-  for (auto input : laGeneric.inputs()) {
-    AffineMap resMap = laGeneric.getIndexingMapsArray()[ioCount++];
-    Value rootOp;
-    if (foldViewLikeOperands(rewriter, input, resMap, rootOp,
-                             toBeErasedViewLikeOps)
-            .failed()) {
-      return failure();
-    }
-    if (rootOp != input)
-      changedIOCount++;
-    newInputs.push_back(rootOp);
-    newMaps.push_back(resMap);
-  }
-  SmallVector<Value, 4U> newOutputs;
-  for (auto output : laGeneric.outputs()) {
-    AffineMap resMap = laGeneric.getIndexingMapsArray()[ioCount++];
-    Value rootOp;
-    if (foldViewLikeOperands(rewriter, output, resMap, rootOp,
-                             toBeErasedViewLikeOps)
-            .failed()) {
-      return failure();
-    }
-    if (rootOp != output)
-      changedIOCount++;
-    newOutputs.push_back(rootOp);
-    newMaps.push_back(resMap);
-  }
-  if (changedIOCount == 0)
-    return failure();
-
-  SmallVector<StringRef> iteratorTypes = llvm::to_vector<4>(
-      laGeneric.getIteratorTypes().getAsValueRange<StringAttr>());
-  auto newLaGenericOp = rewriter.create<linalg::GenericOp>(
-      loc, newInputs, newOutputs, newMaps, iteratorTypes);
-  rewriter.inlineRegionBefore(laGeneric->getRegion(0),
-                              newLaGenericOp.getRegion(),
-                              newLaGenericOp.getRegion().begin());
-  rewriter.replaceOp(laGeneric, newLaGenericOp.getResults());
-  for (auto op : toBeErasedViewLikeOps) {
-    rewriter.eraseOp(op);
-  }
-  return success();
-}
-
 struct MILARewritePattern : public OpRewritePattern<linalg::GenericOp> {
   using OpRewritePattern<linalg::GenericOp>::OpRewritePattern;
 
@@ -260,143 +67,6 @@ struct MILARewritePattern : public OpRewritePattern<linalg::GenericOp> {
                                 PatternRewriter &b) const override;
 };
 } // end anonymous namespace
-
-// This function will create a permutation that will permute the originalMap to
-// be a MinorIdentityWithBroadcast. This is used to add a permutation later in
-// the chain.
-// e.g. :
-// (d0, d1, d2, d4) -> (0, d1) was the map
-// This function will return a permutation : [0, 3, 2, 1] s.t.
-// apply it to the original map would result in
-// (d0, d4, d2, d1) -> (0, d1) in effect.
-static void createPermutationForMinorIdentityWithBroadcast(
-    const AffineMap &originalMap, SmallVectorImpl<uint32_t> &perm) {
-  for (uint32_t i = 0; i < originalMap.getNumInputs(); ++i) {
-    perm.push_back(i);
-  }
-
-  llvm::SmallSet<uint32_t, 4> foundInputDims;
-  for (const auto &idxAndValue : llvm::enumerate(originalMap.getResults())) {
-    auto idx = idxAndValue.index();
-    AffineExpr resultExpr = idxAndValue.value();
-    if (resultExpr.isa<AffineDimExpr>()) {
-      foundInputDims.insert(originalMap.getDimPosition(idx));
-    }
-  }
-
-  for (const auto &idxAndValue : llvm::enumerate(originalMap.getResults())) {
-    auto idx = idxAndValue.index();
-    AffineExpr resultExpr = idxAndValue.value();
-    if (resultExpr.isa<AffineDimExpr>()) {
-      auto swap1 = originalMap.getDimPosition(idx);
-      auto swap2 =
-          originalMap.getNumInputs() - originalMap.getNumResults() + idx;
-      perm[swap1] = swap2;
-      // Only do swap if the output expr does not define another place for the
-      // other input dim
-      if (!foundInputDims.contains(swap2)) {
-        perm[swap2] = swap1;
-      }
-    }
-  }
-}
-
-// This function will take a input Value and a index map that represents the
-// coordinate mapping that could be a combination of tranposes and broadcasts
-// and insert the necessary TransformOps
-static Value insertTransposeAndBroadcastTransforms(PatternRewriter &b,
-                                                   MemRefType outType,
-                                                   Value inp,
-                                                   AffineMap inpIdxMap) {
-  if (!inpIdxMap.isIdentity()) {
-    Location loc = inp.getLoc();
-    auto inpType = inp.getType().template cast<MemRefType>();
-    ArrayRef<int64_t> inpShape = inpType.getShape();
-    ArrayRef<int64_t> outShape = outType.getShape();
-
-    int64_t diff = outShape.size() - inpShape.size();
-    LLVM_DEBUG(llvm::dbgs() << "Reached makeBroadcast with map " << inpIdxMap
-                            << " and diff = " << diff << "\n");
-
-    SmallVector<uint32_t> bcastDims;
-    SmallVector<uint32_t> bcastInDims;
-    SmallVector<uint32_t> passThroughInDims;
-    SmallVector<uint32_t> perm;
-    createPermutationForMinorIdentityWithBroadcast(inpIdxMap, perm);
-    auto permMap = AffineMap::getPermutationMap(perm, b.getContext());
-    inpIdxMap = inpIdxMap.compose(permMap);
-    assert(
-        (inpIdxMap.isMinorIdentityWithBroadcasting(&bcastDims)) &&
-        "this is guranteed by createPermutationForMinorIdentityWithBroadcast");
-
-    // Broadcast those dimensions that the original linalg.generic map specifies
-    // are broadcast and collect their locations, accounting for the leading
-    // dimensions not represented in that map but which are present in the gemm
-    // coordinates
-    BottomUpTMBuilder bcastTransform(b, inpShape, loc);
-    bool hasBcast = false;
-    for (uint32_t i = 0; i < inpShape.size(); ++i) {
-      if (!llvm::is_contained(bcastDims, i)) {
-        // Here the diff correspond to leading dropped dimensions when going
-        // from output co-ordinates to input co-ordinates.
-        assert(inpIdxMap.getDimPosition(i) == diff + i);
-        passThroughInDims.push_back(diff + i);
-        bcastTransform.passThrough({i}, {i});
-      } else {
-        hasBcast = true;
-        bcastInDims.push_back(diff + i);
-        bcastTransform.broadcast({i}, {outShape[diff + i]});
-      }
-    }
-    if (hasBcast) {
-      inp = b.create<TransformOp>(loc, inp, bcastTransform.get());
-    }
-
-    // Then, add dimensions that are present in the writeback coordinates but
-    // are not present in the additional fusion argument with matching sizes.
-    // This, combined with the previous step, ensures that the view of the
-    // fusion argument has the same dimensions as the gemm output, though they
-    // are not necessarily in the same order.
-    bool isDimAdded = false;
-    BottomUpTMBuilder addDimtransform(
-        b, inp.getType().cast<ShapedType>().getShape(), loc);
-    for (uint32_t i = 0; i < outShape.size(); ++i) {
-      unsigned int startIdx = i - diff;
-      if (llvm::is_contained(bcastInDims, i)) {
-        addDimtransform.passThrough({i}, {startIdx});
-      } else if (llvm::is_contained(passThroughInDims, i)) {
-        addDimtransform.passThrough({i}, {startIdx});
-      } else {
-        isDimAdded = true;
-        SmallString<8> name;
-        ("exp" + Twine(i)).toVector(name);
-        addDimtransform.addDim(name, i, outShape[perm[i]]);
-      }
-    }
-    if (isDimAdded) {
-      inp = b.create<TransformOp>(loc, inp, addDimtransform.get());
-    }
-
-    // Permute the dimensions of the fusion argument to match those of the gemm
-    // writeback by applying the inverse of the permutation that would have made
-    // the original indexing map into a minor identity with broadcast. The
-    // inverse of that permutation takes the gemm writeback coordinates and
-    // scatters them into positions that match the non-identity indexing pattern
-    // of the fusion argument.
-    if (!permMap.isIdentity()) {
-      BottomUpTMBuilder permtransform(
-          b, inp.getType().cast<ShapedType>().getShape(), loc);
-      llvm::SmallVector<uint32_t, 4> identityVec;
-      for (uint32_t i = 0; i < outShape.size(); ++i) {
-        identityVec.push_back(i);
-      }
-      permtransform.passThrough(identityVec, perm);
-      inp = b.create<TransformOp>(loc, inp, permtransform.get());
-    }
-    return inp;
-  }
-  return inp;
-}
 
 static void insertLoadFromOtherSource(PatternRewriter &b, Location loc,
                                       GlobalStoreOp gemmStoreOp, Value srcOp,
@@ -498,7 +168,8 @@ Value applyTransforms(PatternRewriter &b, GlobalStoreOp gemmStoreOp, Value inp,
 
   // 1. insert broadcast op if necessary
   MemRefType outType = gemmStoreOp.getDest().getType();
-  ret = insertTransposeAndBroadcastTransforms(b, outType, ret, outToInpMap);
+  ret = insertTransposeAndBroadcastTransforms(b, outType.getShape(), ret,
+                                              outToInpMap);
 
   // 2. also create global_store from global to regs
   //    TODO(sjw): make sure output buffer writes (means these inputs will be
@@ -765,7 +436,6 @@ LogicalResult MILARewritePattern::matchAndRewrite(linalg::GenericOp laGeneric,
 void RockLinalgAlignPass::runOnOperation() {
   MLIRContext *ctx = &getContext();
   RewritePatternSet patterns(ctx);
-  patterns.add<InlineViewLikeOperandsLinalgRewritePattern>(ctx);
   patterns.add<MILARewritePattern>(ctx);
   if (failed(applyPatternsAndFoldGreedily(getOperation(), std::move(patterns))))
     signalPassFailure();
