+++ mlir/lib/Dialect/Rock/Transforms/FoldTranspose.cpp
@@ -21,9 +21,11 @@
 #include "mlir/Dialect/Rock/IR/RockGemmWrapperInterface.h"
 #include "mlir/Dialect/Rock/IR/TransformMapBuilder.h"
 #include "mlir/Dialect/Rock/Passes.h"
+#include "mlir/Dialect/Rock/utility/transformMapUtils.h"
 
 #include "mlir/Transforms/GreedyPatternRewriteDriver.h"
 
+#include "llvm/ADT/SmallSet.h"
 #include "llvm/Support/Debug.h"
 
 namespace mlir {
@@ -39,6 +41,199 @@ using namespace mlir;
 using namespace mlir::rock;
 
 namespace {
+
+// This rewrite will rewrite the linalg IO that has view like-ops surrounding
+// them to be consumed by the linalg operation itself adjusting the indexing
+// maps to faithfully represent them.
+struct InlineViewLikeOperandsLinalgRewritePattern
+    : public OpRewritePattern<linalg::GenericOp> {
+  using OpRewritePattern<linalg::GenericOp>::OpRewritePattern;
+
+  LogicalResult matchAndRewrite(linalg::GenericOp laGeneric,
+                                PatternRewriter &rewriter) const override;
+};
+
+// This function will take strides calculated using reassociation of dimensions
+// that are modified by view-like ops and recompute the affine expr from src
+// view to dest view.
+static void getDelinearizedAffineExpr(mlir::ArrayRef<int64_t> strides,
+                                      mlir::ArrayRef<int64_t> shapes,
+                                      Builder &b, unsigned int position,
+                                      SmallVectorImpl<mlir::AffineExpr> &res) {
+  AffineExpr resultExpr = b.getAffineDimExpr(position);
+  int64_t rank = strides.size();
+  // If the rank is 1, expand or collapse shapes will just
+  // pass-through the dimensions.
+  if (rank == 1) {
+    res[0] = resultExpr;
+    return;
+  }
+  for (int i = 0; i < rank; i++) {
+    // If the current shape is 1 and the rank is non-zero,
+    // could only mean it is being broadcasted. Hence,
+    // putting zero.
+    if (shapes[i] == 1) {
+      res[i] = resultExpr * 0;
+    } else {
+      // Recording the vector offsets here.
+      res[i] = resultExpr;
+      // There is no point of putting a modulo if the size
+      // is equivalent to that.
+      if (i - 1 >= 0 && shapes[i] != strides[i - 1]) {
+        res[i] = res[i] % strides[i - 1];
+      }
+
+      if (shapes[i] > strides[i]) {
+        // We only need the floorDiv if the dimSize
+        // is larger than the stride
+        res[i] = res[i].floorDiv(strides[i]);
+      } else if (shapes[i] < strides[i]) {
+        // if the shape is smaller than the stride
+        // expr might as well be zero.
+        res[i] = res[i] * 0;
+      }
+
+      // The resultExpr has to propagated anyway for
+      // other dimensions where the recording in the above
+      // will do the neccesary checks to remove the modulo
+      if (i - 1 >= 0) {
+        resultExpr = resultExpr % strides[i - 1];
+      }
+      resultExpr = resultExpr.floorDiv(strides[i]);
+    }
+  }
+  return;
+}
+
+// This function will create a affine map that represent the mapping
+// from higher rank memref type to lower rank memref type.
+static AffineMap createHigherToLowerRankViewAffineMap(
+    PatternRewriter &rewriter,
+    ArrayRef<ReassociationIndices> reassociationIndices,
+    const MemRefType &higherRankType, const MemRefType &lowerRankType) {
+  SmallVector<AffineExpr, 4> resultExprs;
+  int iDimCount = 0;
+  for (const SmallVector<int64_t, 2> &groups : reassociationIndices) {
+    assert(!groups.empty() && "association indices groups cannot be empty");
+    unsigned groupSize = groups.size();
+    SmallVector<int64_t> suffixProduct(groupSize);
+    // Calculate suffix product for all collapse op source dimension sizes.
+    suffixProduct[groupSize - 1] = 1;
+    for (unsigned i = groupSize - 1; i > 0; i--)
+      suffixProduct[i - 1] =
+          suffixProduct[i] * higherRankType.getDimSize(groups[i]);
+    SmallVector<int64_t> shapes(groupSize);
+    for (unsigned i = 0; i < groupSize; i++) {
+      shapes[i] = higherRankType.getDimSize(groups[i]);
+    }
+    // Derive the index values along all dimensions of the source
+    // corresponding to the index wrt to collapsed shape op output.
+    SmallVector<AffineExpr, 4> srcIndexExpr(suffixProduct.size());
+    getDelinearizedAffineExpr(suffixProduct, shapes, rewriter, iDimCount++,
+                              srcIndexExpr);
+    for (unsigned i = 0; i < groupSize; i++) {
+      resultExprs.push_back(srcIndexExpr[i]);
+    }
+  }
+  auto representativeMap = AffineMap::get(
+      /*numDims=*/lowerRankType.cast<ShapedType>().getShape().size(),
+      /*numSymbols=*/0, resultExprs, rewriter.getContext());
+  return representativeMap;
+}
+
+// This function will traverse the operands of the linalg.generic folding
+// the view like ops to the indexing maps and returning the ultimate
+// folded map as well as the root operand.
+LogicalResult
+foldViewLikeOperands(PatternRewriter &rewriter, Value op, AffineMap &foldedMap,
+                     Value &rootOp,
+                     SmallVectorImpl<Operation *> &toBeErasedViewLikeOps) {
+  if (memref::CollapseShapeOp collapseOp =
+          op.getDefiningOp<memref::CollapseShapeOp>()) {
+    SmallVector<mlir::ReassociationIndices, 4U> reassociationIndices =
+        collapseOp.getReassociationIndices();
+    MemRefType lowerRankType = collapseOp.getType();
+    MemRefType higherRankType = collapseOp.getSrcType();
+    auto representativeMap = createHigherToLowerRankViewAffineMap(
+        rewriter, reassociationIndices, higherRankType, lowerRankType);
+    foldedMap = representativeMap.compose(foldedMap);
+    toBeErasedViewLikeOps.push_back(collapseOp);
+    return foldViewLikeOperands(rewriter, collapseOp.getViewSource(), foldedMap,
+                                rootOp, toBeErasedViewLikeOps);
+  }
+  if (memref::ExpandShapeOp expandOp =
+          op.getDefiningOp<memref::ExpandShapeOp>()) {
+    SmallVector<mlir::ReassociationIndices, 4U> reassociationIndices =
+        expandOp.getReassociationIndices();
+    MemRefType higherRankType = expandOp.getType();
+    MemRefType lowerRankType = expandOp.getSrcType();
+    auto representativeMap = createHigherToLowerRankViewAffineMap(
+        rewriter, reassociationIndices, higherRankType, lowerRankType);
+    // We take the inverse here because in expand shape it is going from lower
+    // to higher rank.
+    representativeMap = inversePermutation(representativeMap);
+    foldedMap = representativeMap.compose(foldedMap);
+    toBeErasedViewLikeOps.push_back(expandOp);
+    return foldViewLikeOperands(rewriter, expandOp.getViewSource(), foldedMap,
+                                rootOp, toBeErasedViewLikeOps);
+  }
+  rootOp = op;
+  return success();
+}
+
+LogicalResult InlineViewLikeOperandsLinalgRewritePattern::matchAndRewrite(
+    linalg::GenericOp laGeneric, PatternRewriter &rewriter) const {
+  Location loc = laGeneric.getLoc();
+  int ioCount = 0;
+
+  SmallVector<AffineMap, 4U> newMaps;
+  SmallVector<Value, 4U> newInputs;
+  unsigned int changedIOCount = 0;
+
+  SmallVector<Operation *> toBeErasedViewLikeOps;
+  for (auto input : laGeneric.inputs()) {
+    AffineMap resMap = laGeneric.getIndexingMapsArray()[ioCount++];
+    Value rootOp;
+    if (foldViewLikeOperands(rewriter, input, resMap, rootOp,
+                             toBeErasedViewLikeOps)
+            .failed()) {
+      return failure();
+    }
+    if (rootOp != input)
+      changedIOCount++;
+    newInputs.push_back(rootOp);
+    newMaps.push_back(resMap);
+  }
+  SmallVector<Value, 4U> newOutputs;
+  for (auto output : laGeneric.outputs()) {
+    AffineMap resMap = laGeneric.getIndexingMapsArray()[ioCount++];
+    Value rootOp;
+    if (foldViewLikeOperands(rewriter, output, resMap, rootOp,
+                             toBeErasedViewLikeOps)
+            .failed()) {
+      return failure();
+    }
+    if (rootOp != output)
+      changedIOCount++;
+    newOutputs.push_back(rootOp);
+    newMaps.push_back(resMap);
+  }
+  if (changedIOCount == 0)
+    return failure();
+
+  SmallVector<StringRef> iteratorTypes = llvm::to_vector<4>(
+      laGeneric.getIteratorTypes().getAsValueRange<StringAttr>());
+  auto newLaGenericOp = rewriter.create<linalg::GenericOp>(
+      loc, newInputs, newOutputs, newMaps, iteratorTypes);
+  rewriter.inlineRegionBefore(laGeneric->getRegion(0),
+                              newLaGenericOp.getRegion(),
+                              newLaGenericOp.getRegion().begin());
+  rewriter.replaceOp(laGeneric, newLaGenericOp.getResults());
+  for (auto op : toBeErasedViewLikeOps) {
+    rewriter.eraseOp(op);
+  }
+  return success();
+}
 struct RockFoldTransposePass
     : public rock::impl::RockFoldTransposePassBase<RockFoldTransposePass> {
   void runOnOperation() override;
@@ -135,27 +330,32 @@ struct RemoveTrivialTransposePattern
   }
 };
 
-/// If the input to a linalg.generic is the output of a conv2d and the indexing
-/// map for that input is a non-trivial permutation of an identity, convert that
-/// indexing map to a transpose. This must happen before gridwise gemm
-/// conversion because all the transforms on the convolution output are
-/// collected at that time.
-struct FoldTransposingConvAccess : OpRewritePattern<linalg::GenericOp> {
+// If the input to a linalg.generic is the output of a rock compute op and the
+// indexing map for that input is a non-trivial permutation of an identity,
+// convert that indexing map to a transpose, expansions and/or collapses. This
+// must happen before gridwise gemm conversion because all the transforms on the
+// rock compute op output are collected at that time.
+struct FoldRockOutputTransforms : OpRewritePattern<linalg::GenericOp> {
   using OpRewritePattern<linalg::GenericOp>::OpRewritePattern;
 
   LogicalResult matchAndRewrite(linalg::GenericOp laGeneric,
                                 PatternRewriter &b) const override {
-    Location loc = laGeneric->getLoc();
     // We do this out-of-line so as to not invalidate our iterator
-    SmallVector<
-        std::tuple<unsigned, Value, AffineMap, SmallVector<uint32_t, 4>>>
-        toReplace;
+    SmallVector<std::tuple<unsigned, Value, AffineMap>> toReplace;
+
+    auto laGenericOut = laGeneric.getOutputOperand(0);
+    auto laGenericOutIdxMap = laGeneric.getTiedIndexingMap(laGenericOut);
 
     for (OpOperand &operand : laGeneric->getOpOperands()) {
       Value opValue = operand.get();
       Operation *convUser = getGemmWrapperUser(opValue);
       if (!convUser)
         continue;
+      // If the laGenericOut and rockOp user has the same type
+      // there is nothing to be done here.
+      if (operand.get().getType() == laGenericOut->get().getType()) {
+        continue;
+      }
 
       for (Operation *user : opValue.getUsers()) {
         if (isa<linalg::GenericOp>(user) && user != laGeneric) {
@@ -167,64 +367,52 @@ struct FoldTransposingConvAccess : OpRewritePattern<linalg::GenericOp> {
       if (!isa_and_nonnull<memref::AllocOp>(opValue.getDefiningOp()))
         continue;
 
-      AffineMap idxMap = laGeneric.getTiedIndexingMap(&operand);
-      if (idxMap.isMinorIdentityWithBroadcasting())
-        continue;
+      AffineMap inpIdxMap = laGeneric.getTiedIndexingMap(&operand);
+      auto invertInpIdxMap = inversePermutation(inpIdxMap);
+      auto inToOutIdxMap = laGenericOutIdxMap.compose(invertInpIdxMap);
       SmallVector<uint32_t, 4> permutation;
-      if (!idxMap.isPermutationOfMinorIdentityWithBroadcasting(permutation))
+      if (!inToOutIdxMap.isPermutationOfMinorIdentityWithBroadcasting(
+              permutation))
         continue;
 
       unsigned opIndex = operand.getOperandNumber();
-      toReplace.emplace_back(opIndex, opValue, idxMap, permutation);
+      toReplace.emplace_back(opIndex, opValue, inToOutIdxMap);
     }
 
     // Actually do the rewrites, if any
     SmallVector<AffineMap> idxMaps = laGeneric.getIndexingMapsArray();
+    SmallVector<Value> inputs = laGeneric.inputs();
     for (auto &tuple : toReplace) {
       unsigned opIndex = std::get<0>(tuple);
       Value opValue = std::get<1>(tuple);
       AffineMap idxMap = std::get<2>(tuple);
-      SmallVector<uint32_t, 4> permutation = std::get<3>(tuple);
-      LLVM_DEBUG(llvm::dbgs() << "Replacing index map with permutation ");
-      LLVM_DEBUG(llvm::interleaveComma(permutation, llvm::dbgs()));
-      LLVM_DEBUG(llvm::dbgs() << "\n");
-
       auto allocation = cast<memref::AllocOp>(opValue.getDefiningOp());
       // Swap out the allocation for the form it needs to take in order to
       // eliminate the non-trivial map.
-      ArrayRef<int64_t> shape = opValue.getType().cast<MemRefType>().getShape();
-      SmallVector<int64_t, 4> newShape(shape.size(), -1LL);
-      SmallVector<uint32_t, 4> endIdentity;
-      for (uint32_t i = 0, e = shape.size(); i < e; ++i) {
-        endIdentity.push_back(i);
-        newShape[permutation[i]] = shape[i];
-      }
+      auto newShape =
+          laGenericOut->get().getType().cast<ShapedType>().getShape();
 
       // All this new stuff needs to go where the old memref.alloc was
       PatternRewriter::InsertionGuard guard(b);
       b.setInsertionPointAfterValue(allocation);
-      auto newAllocType = allocation.getType()
-                              .cast<MemRefType>()
-                              .clone(newShape)
-                              .cast<MemRefType>();
+      auto newAllocType =
+          MemRefType::get(newShape, allocation.getType().getElementType());
+      MemRefType outType = allocation.getType();
       Value newAlloc =
-          b.replaceOpWithNewOp<memref::AllocOp>(allocation, newAllocType);
+          b.create<memref::AllocOp>(allocation.getLoc(), newAllocType);
+      Value transformedNewAlloc = insertTransposeAndBroadcastTransforms(
+          b, outType.getShape(), newAlloc, idxMap);
+      llvm::SmallPtrSet<Operation *, 2> skips = {
+          laGeneric, transformedNewAlloc.getDefiningOp()};
+      opValue.replaceAllUsesExcept(transformedNewAlloc, skips);
 
-      rock::BottomUpTMBuilder permuteMapBuilder(b, newShape, loc);
-      permuteMapBuilder.passThrough(endIdentity, permutation);
-      TransformMapAttr permuteMapAttr = permuteMapBuilder.get();
-      auto permuted =
-          b.create<rock::TransformOp>(loc, newAlloc, permuteMapAttr);
-      llvm::SmallPtrSet<Operation *, 2> skips = {laGeneric, permuted};
-      newAlloc.replaceAllUsesExcept(permuted, skips);
-
-      // Correct indexing maps
-      AffineMap composed =
-          idxMap.compose(permuteMapAttr.getMap().getAffineMap());
-      idxMaps[opIndex] = composed;
+      // Correct indexing maps and changing the inputs
+      idxMaps[opIndex] = laGenericOutIdxMap;
+      inputs[opIndex] = newAlloc;
     }
     laGeneric->setAttr(laGeneric.indexing_mapsAttrName(),
                        b.getAffineMapArrayAttr(idxMaps));
+    laGeneric.getInputsMutable().assign(inputs);
     return success(!toReplace.empty());
   }
 };
@@ -238,7 +426,9 @@ void RockFoldTransposePass::runOnOperation() {
   }
 
   RewritePatternSet patternsTP(ctx);
-  patternsTP.add<RemoveTrivialTransposePattern, FoldTransposingConvAccess>(ctx);
+  patternsTP.add<RemoveTrivialTransposePattern,
+                 InlineViewLikeOperandsLinalgRewritePattern,
+                 FoldRockOutputTransforms>(ctx);
   if (failed(
           applyPatternsAndFoldGreedily(getOperation(), std::move(patternsTP))))
     signalPassFailure();
